requirements.txt
fastapi
uvicorn[standard]
databases
asyncpg
python-dotenv
pydantic[email]
python-jose[cryptography]
bcrypt
openai>=1.3.5,<2.0.0
google-analytics-data
google-auth
requests
tenacity
apscheduler
sqlalchemy
psycopg2-binary
gunicorn
aiofiles
python-multipart
numpy
pyjwt
jinja2
msal>=1.22.0
slowapi 
pip-audit
weasyprint
requests_oauthlib
python-docx